style: CountDisplay component #48
